pyproject.toml
[tool.poetry]
name = "fig"
version = "0.1.0"
description = "Some cheeky pun about configuration and the fruits of your labour"
authors = ["Julian True <julian.allan.true@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "*"
toml = "*"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
